// frontend/src/app/api/generated/model/retail-product.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RetailProduct { 
    id: string;
    code: string;
    name: string;
    defaultPrice: number;
    active: boolean;
}